Update Uploading-Files-to-your-Reclaim-Hosting-Account.md
Mostly small with two exceptions. I added in a link to the FTP documentation at the end of the FTP section and, as per the FTP documentation, I added a note to the first SFTP mention to clarify that ti will only work for the default cPanel user, not a separate FTP user account.